Shorten duration
A shorted duration could prevent burnout or tiredness from a too long event. A weekend is a very common duration for a lot of hackathons/game-jams. This way people that have a day job could also attend the full jam, and it requires less commitment from participants, broadening the participation. This request pairs well with request https://github.com/H-R-B-Y/game_jam_proposal/pull/1 (unrestricted language/library usage), as the shorter duration could be a problem to some people if only allowed to use C/C++

I quite like the idea of a 5 day jam _a la_ primeagen game in a week series, could be considered for a future jam if this one is successful.